RS - Rev.1 - Функциональность Event. Добавить контроллер, сервисы, репозитории, исправить, добавить клиент статистики для Public и Admin (3/3). Требует доработки

+++ main-service/src/main/java/ru/practicum/main/dto/request/event/SearchOfEventByAdminDto.java
@@ -13,9 +13,9 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 public class SearchOfEventByAdminDto {
-    private List<Integer> users;
+    private List<Long> users;
     private List<String> states;
-    private List<Integer> categories;
+    private List<Long> categories;
     private LocalDateTime rangeStart;
     private LocalDateTime rangeEnd;
 }
